refactor(mobile): split android manifest helpers
